DI: refactored StatePensionServiceSpec into 3 files
