+++ test/uk/gov/hmrc/statepension/services/CitizenDetailsServiceSpec.scala
@@ -18,15 +18,17 @@ package uk.gov.hmrc.statepension.services
 
 import org.mockito.Matchers
 import org.mockito.Mockito.when
+import org.scalatestplus.mockito.MockitoSugar
 import play.api.http.Status.{LOCKED, OK}
 import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.statepension.connectors.CitizenDetailsConnector
 
 import scala.concurrent.Future
 import uk.gov.hmrc.http.HeaderCarrier
+import uk.gov.hmrc.statepension.StatePensionUnitSpec
 
 //TODO change to proper extensions
-class CitizenDetailsServiceSpec extends StatePensionServiceSpec {
+class CitizenDetailsServiceSpec extends StatePensionUnitSpec with MockitoSugar {
 
   val nino: Nino = generateNino()
   val mockCitizenDetailsConnector: CitizenDetailsConnector = mock[CitizenDetailsConnector]
